requirements.txt
boto3==1.17.105
pyyaml==5.4.1
